Started work on the GUI
Working on the GUI is the first step in the project, currently I'm one lesson ahead of schedule on my gantt chart, which means I'm working at a good pace and I am able to cover for myself incase I miss a lesson.

The first steps of the UI include an Entry that will allow the player to search for things, and setting up a title,
I've also started work on setting up a frame to list all the watched movies.