Further work on GUI
Gui has had more additions, and have started a little bit of the backend that is required to have the GUI load. This is mainly the watchedmovies system, which is currently using a moviestorage class that will be read from a file and implemented, this will store the title, cover and id of the movie.
This was all so that the tiling storage could work within the UI.
This has been tested so far in that the program now has a scrollable frame that needs to become horizontally scrolling (currently vertically), that stores the movies and currently (as no movies have been stored / no storage has been implemented) displays an empty "+" icon image, though the image looks horrible at the moment as it was not designed for the current display resolution (will be edited in photoshop at a later date.