J'ai presque fini faire le programme :)
I'm nearly done the program.
This update comes with a major overhaul of alot of what I've worked on.
I fixed the tiling system for the movies on the main page, as the math was a little broken (It had to do with weird interactions between range and enumeration).
I implemented a direct Error Code System that will help later into the maintenance section.
I implemented the API which was given its own file (apiinteractions) as required.
I implemented saving which means the system can now actually be useable, alongside the newly implemented loading, it resaves these every single time an interaction is made, whilst this is taxing its an important solution as CustomTkinter and Tkinter don't allow you to run functions on exit (as far as I know) and it helps prevent data corruption from crashes.
I also implemented the movie preview page, This just gets the movie from the database and uses its dictionary to go through and display everything. It also has the big match case (pythons casewhere) for converting genre ids into genres.
I've also created an error window that displays errors to the user (This is alongside the previous mentioned code outputting system, that shows me the errorcode, python version, and operating system of the user).
Lastly I've added interactions to the home screen, where you can directly view the movies preview page and remove movies from your watchlist.

What's Next?
I said I've nearly finished making the program as there are a few finishing touches I'd like to add but are not necessary. Firstly, I'd be interested into adding some form of multithreading support for the cover images, as its causing the load time of the program to be O(n), just to see the main window, which is suboptimal especially in the case of power users.
Other than that, the icon to add new slots could do with a cleanup, (Where it says + Search to add movies), currently it looks like it does as its just a random image I ripped from the internet.
I could also do with some indication of loading (Which shouldn't be necessary if I manage to implement multi-threading).
And A logo for the app would be nice if I have the spare time.

In Short:
Worked on the program and did all the bulk work, as well as a few details.
Not much to do next, just those last finishing touches and maybe multithreading.
Au Revoir

Note: If you use the program as it currently is, it won't work as I had to hide the API key so it wouldn't get scraped.