Multithreading Support and Cleaned up Icons
Here I added mutlithreading support,
My previous attempts of multithreading had various issues, but after further research I discovered its because Tkinter is not thread safe by default. As such they implemented an "after_idle" function that means as soon as the app.mainloop() is idle, the code added will be run.
Using this we can queue events onto tkinters main loop whilst not being on the main thread.
Then all that I had to do was create the threads for it to run on, parse the functions an arguments, and let it run.
Thanks to load times, I've had to go through and make all of the web images have a backup (unloaded version), so I made 3 images, One for the Add a new image, another for unloaded covers, and another for unloaded backdrops.

There isn't much left to add, maybe a logo? But all the things that I wanted to add after the last commit have been added, and the program is nearing its final state.