Search Window
Added a search window, this is because some movies can't be found as there are bigger movies with the exact same names, or extremely similar names, and just giving the first result doesn't work.

I also added a rating watched in this, So it should give no movies that are rated R in Australia or the US (Only goes to US if there isn't an Australian rating) or the movie isn't rated in the US or Australia.
This is to stop us from getting inappropriate results.

What's next? A help window as required on the assessment task is all thats left.