Help Window
Here I added the help window, that is the final step of the assessment task for the most part.

It has some basic images on a scrolling frame that explains it.

This is a good change that has great impact on the useability of the program, as it no longer relies on the user being an expert, and guides them through the program using images and text.

There is nothing left to do program wise.

EDIT: Sir told me to make a report bug window. That is what I shall now do.