Report Function
In this updated I added a report button and Report form, (which is then shredded upon use and not stored).

This has little to no impact on the actual value of the program. It mainly effects the outlook of the program on how cool it is B)

Next to implement:
Thinking of adding watch providers (netflix etc...)