Watch Providers
In this update I've added the watch providers to the program,
This means within the preview window of movies we can now see where we can buy or stream the movie.

This has an incredible impact on the functionality of the program, and allows users to have it as a one* stop shop (*except for watching the movies). It means users have an actual reason to use the program besides just a list (it now has more functionality then a fancy notepad).

Next to add:
No clue at the moment,
Whatever sir comes up with randomly next lesson.